+++ medico/eliminarMedico.php
@@ -13,6 +13,22 @@ if (!empty($_GET['codMedico']) ){
 
     $codMedico = $_GET['codMedico'];
 
+    $querySearchDni = "SELECT * FROM medicos_hospitales_servicios WHERE cod_medicos  = '$codMedico'";
+    $searchDni = $conexion->query($querySearchDni)->rowCount();
+
+    if($searchDni > 0){
+        $mensaje = [
+            'mensaje' => 'ESTE MEDICO NO LO PUEDES ELIMINAR SI TIENE ASIGNADO UNA HOSPITAL SERVICIO',
+            'alerta' => 'danger'
+        ];
+
+        $_SESSION['mensaje'] = $mensaje;
+        header('Location: listaMedico.php');
+
+        exit();
+    }
+
+
     $query = "DELETE FROM medicos WHERE cod_medicos = '$codMedico'";
     $eliminar = $conexion->query($query);
